move around last commands
